+++ src/RcppExports.cpp
@@ -5,21 +5,22 @@
 
 using namespace Rcpp;
 
-// categorize
-IntegerVector categorize(NumericVector x, NumericVector breaks);
-RcppExport SEXP _santoku_categorize(SEXP xSEXP, SEXP breaksSEXP) {
+// categorize_impl
+IntegerVector categorize_impl(NumericVector x, NumericVector breaks, LogicalVector left);
+RcppExport SEXP _santoku_categorize_impl(SEXP xSEXP, SEXP breaksSEXP, SEXP leftSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericVector >::type x(xSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type breaks(breaksSEXP);
-    rcpp_result_gen = Rcpp::wrap(categorize(x, breaks));
+    Rcpp::traits::input_parameter< LogicalVector >::type left(leftSEXP);
+    rcpp_result_gen = Rcpp::wrap(categorize_impl(x, breaks, left));
     return rcpp_result_gen;
 END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_santoku_categorize", (DL_FUNC) &_santoku_categorize, 2},
+    {"_santoku_categorize_impl", (DL_FUNC) &_santoku_categorize_impl, 3},
     {NULL, NULL, 0}
 };