+++ src/categorize.cpp
@@ -2,10 +2,9 @@
 using namespace Rcpp;
 
 // [[Rcpp::export]]
-IntegerVector categorize(NumericVector x, NumericVector breaks) {
+IntegerVector categorize_impl(NumericVector x, NumericVector breaks, LogicalVector left) {
   int xs = x.size();
   int bs = breaks.size();
-  LogicalVector left = breaks.attr("left");
   if (left.size() != bs) Rcpp::stop("`left` of different size to `breaks`");
 
   IntegerVector codes(xs, NA_INTEGER);